* cbase, fix compilation/linking issues on Windows

+++ source/main/include/cbase/private/c_integer_inline.h
@@ -125,12 +125,12 @@ namespace ncore
         inline u64 alignUp(u64 value, u32 alignment)
         {
             ASSERTS(ispo2(alignment) == true, "Error: alignment value should be a power of 2");
-            return ((value + (alignment - 1)) & (~(alignment - 1)));
+            return ((value + (alignment - 1)) & (~((u64)alignment - 1)));
         }
         inline u64 alignUp(u64 value, s32 alignment)
         {
             ASSERTS(ispo2(alignment) == true, "Error: alignment value should be a power of 2");
-            return ((value + ((u32)alignment - 1)) & (~((u32)alignment - 1)));
+            return ((value + ((u64)alignment - 1)) & (~((u64)alignment - 1)));
         }
 
         // Check if value is aligned
